moved rwt payload generator
