+++ src/services/certificate-payload/CertificatePayloadGeneratorRwt.ts
@@ -1,11 +1,11 @@
 import { Inject, Service } from 'typedi';
 import moment from 'moment';
-import { ITestResult } from '../models/ITestResult';
-import { TechRecordsService } from './TechRecordsService';
-import { TEST_RESULTS, VEHICLE_TYPES } from '../models/Enums';
-import { IRoadworthinessCertificateData } from '../models/IRoadworthinessCertificateData';
-import { DefectService } from './DefectService';
-import { ICertificatePayloadGenerator } from './ICertificatePayloadGenerator';
+import { ITestResult } from '../../models/ITestResult';
+import { TechRecordsService } from '../TechRecordsService';
+import { TEST_RESULTS, VEHICLE_TYPES } from '../../models/Enums';
+import { IRoadworthinessCertificateData } from '../../models/IRoadworthinessCertificateData';
+import { DefectService } from '../DefectService';
+import { ICertificatePayloadGenerator } from '../ICertificatePayloadGenerator';
 
 @Service()
 export class CertificatePayloadGeneratorRwt implements ICertificatePayloadGenerator {
